Fix saving logs on CI
