+++ test/cloudtest/pkg/providers/packet/packet_provider.go
@@ -431,7 +431,7 @@ func (pi *packetInstance) updateProject() error {
 
 func (pi *packetInstance) createKey(keyFile string) ([]string, error) {
 	today := time.Now()
-	genID := fmt.Sprintf("%d-%d-%d-%s", today.Year(), today.Month(), today.Day(), shell.NewRandomStr(10))
+	genID := fmt.Sprintf("%d-%d-%d-%s", today.Year(), today.Month(), today.Day(), utils.NewRandomStr(10))
 	pi.keyID = "dev-ci-cloud-" + genID
 
 	out := strings.Builder{}
